Implemented PR feedback, specifically for dividendComputation Javascript proof construction.
Seperated the verification code into it's own seperate file and created a method that computes the challenge, independently of being inside another method.

Deleted redundant methods in helpers.js and made use of more succinct programming patterns for constructing test notes
